feat: move SMTP email code to new email-smtp included plugin
