+++ imports/plugins/core/email/client/index.js
@@ -2,7 +2,4 @@ import "./templates/email.html";
 import "./templates/email.js";
 
 export { default as EmailTableColumn } from "./components/emailTableColumn";
-
-export { default as EmailConfig } from "./containers/emailConfig";
 export { default as EmailLogs } from "./containers/emailLogs";
-export { default as EmailSettings } from "./containers/emailSettings";
