+++ imports/plugins/core/email/server/index.js
@@ -1,6 +1,1 @@
-import { Meteor } from "meteor/meteor";
 import "./i18n";
-import methods from "./methods";
-import "./startup";
-
-Meteor.methods(methods);
