+++ imports/plugins/included/email-smtp/client/index.js
@@ -1,0 +1,2 @@
+import "./containers/SMTPEmailConfigContainer";
+import "./containers/SMTPEmailSettingsContainer";
