+++ imports/plugins/included/email-smtp/server/index.js
@@ -1,0 +1,5 @@
+import { Meteor } from "meteor/meteor";
+import methods from "./methods";
+import "./startup";
+
+Meteor.methods(methods);
