+++ imports/plugins/included/email-smtp/server/methods/verifySettings.js
